requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
prometheus-client==0.19.0
pytest==7.4.3
httpx==0.25.2
asyncpg==0.29.0
aiokafka==0.8.1
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 